+++ tests/test_update_subdivisions.py
@@ -7,6 +7,7 @@ import shutil
 import unittest
 unittest.TestLoader.sortTestMethodsUsing = None
 
+# @unittest.skip("Skipping update_subdivisions unit tests.")
 class Update_Subdivisions_Tests(unittest.TestCase):
     """
     Test suite for testing update_subdivisions.py script that that is used for the