+++ tests/test_utils.py
@@ -4,7 +4,7 @@ import os
 import unittest
 unittest.TestLoader.sortTestMethodsUsing = None
 
-# @unittest.skip("")
+# @unittest.skip("Skipping utils unit tests.")
 class Utils_Tests(unittest.TestCase):
     """
     Test suite for testing the utilities module that consists of various functions